requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.6
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==20.8b1
bs4==0.0.1
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cssselect==1.1.0
Django==3.1.7
django-bootstrap4==3.0.1
django-celery-beat==2.2.0
django-chartjs==2.2.1
django-filter==2.4.0
django-flower==1.0.0
django-jinja==2.7.0
django-mathfilters==1.0.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-timezone-field==4.1.2
fake-useragent==0.1.11
flower==0.9.7
gevent==21.1.2
greenlet==1.1.0
humanize==0.5.1
idna==2.10
jellyfish==0.8.2
Jinja2==3.0.1
joblib==1.0.1
jsonfield==3.1.0
kombu==4.6.11
lxml==4.6.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.20.2
pandas==1.2.4
parse==1.19.0
pathspec==0.8.1
Pillow==8.2.0
plumbum==1.7.0
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pyee==8.1.0
pypiwin32==223
pyppeteer==0.2.5
pyquery==1.4.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
recordlinkage==0.14
redis==3.5.3
regex==2021.3.17
requests==2.25.1
requests-html==0.10.0
rpyc==5.0.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.59.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==1.3.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1
zope.event==4.5.0
zope.interface==5.4.0
